dev(more_networks): Moving RPC stuff
