--- rust/theoros/src/hyperlane/mod.rs
@@ -1,1 +1,0 @@
-pub mod calls;
