+++ rust/theoros/src/rpc/evm/mod.rs
@@ -1,0 +1,3 @@
+pub mod hyperlane;
+
+pub use hyperlane::*;
