+++ rust/theoros/src/rpc/mod.rs
@@ -1,16 +1,2 @@
-pub mod hyperlane;
-pub mod pragma_feeds_registry;
-
-pub use hyperlane::*;
-pub use pragma_feeds_registry::*;
-
-use starknet::providers::{jsonrpc::HttpTransport, JsonRpcClient};
-use url::Url;
-
-pub struct StarknetRpc(JsonRpcClient<HttpTransport>);
-
-impl StarknetRpc {
-    pub fn new(rpc_url: Url) -> Self {
-        Self(JsonRpcClient::new(HttpTransport::new(rpc_url)))
-    }
-}
+pub mod evm;
+pub mod starknet;
