+++ rust/theoros/src/storage/mod.rs
@@ -8,7 +8,7 @@ pub use validator::*;
 
 use starknet::core::types::Felt;
 
-use crate::rpc::{HyperlaneCalls, PragmaFeedsRegistryCalls, StarknetRpc};
+use crate::rpc::starknet::{HyperlaneCalls, PragmaFeedsRegistryCalls, StarknetRpc};
 
 /// Theoros storage that contains:
 ///   * a set of all available feed ids,
